casting collection size to double before dividing

+++ document-similarity/src/main/java/pl/edu/icm/coansys/similarity/documents/function/AvgSimilarityFunction.java
@@ -32,7 +32,7 @@ public class AvgSimilarityFunction implements SimilarityFunction {
             similarity *= keywordsSimilarities.get(i);
         }
         
-        similarity /= keywordsSimilarities.size();
+        similarity /= (double) keywordsSimilarities.size();
         return similarity;
     }
 }
